Move notify utilities, refactor health & add CSV
Refactor project layout: move notify-related modules from app/api/resend to app/utils/notify (resend.py, send_email.py, and package init), and relocate the health endpoint from app/api/health.py to app/utils/health.py. Update app/api/root.py and app/api/routes.py to reflect the new import paths. Add LinkedInConnections.csv under app/api/prospects. Rename tests/test_llm.py to tests/test_prompt.py and apply related test updates (tests/test_routes.py). These changes separate API routes from utility implementations and add prospect data.
