+++ app/utils/notify/resend.py
@@ -4,7 +4,7 @@ import os
 from app.utils.make_meta import make_meta
 from fastapi import APIRouter, Query, Path, Body, HTTPException
 from app.utils.db import get_db_connection
-from app.utils.send_email import send_email_resend
+from app.utils.notify.send_email import send_email_resend
 
 router = APIRouter()
 base_url = os.getenv("BASE_URL", "http://localhost:8000")
